Use SQLiteDatabaseHelper to save Log entries to db. is also capable of saving and fetching all tasks. But not able to show list of tasks in UI, nor edit tasks in UI. Need to implement Task manipulation (GUI).

+++ MyTimer/app/src/main/java/fi/fimurito/mytimer/data/model/TaskLog.kt
@@ -12,8 +12,8 @@ import java.time.LocalDateTime
 @Entity(tableName = "tasklogs")
 @TypeConverters(MyTypeConverters::class)
 class TaskLog(
-    @PrimaryKey val id: Long? = null,
-    @ColumnInfo val taskId: Long? = null,
+    @PrimaryKey val id: Int? = null,
+    @ColumnInfo val taskId: Int? = null,
     @ColumnInfo val beginDate: LocalDateTime?,
     @ColumnInfo val endDate: LocalDateTime?,
     @ColumnInfo val comment: String? = null,
